+++ CHANGELOG.md
@@ -6,9 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## `Unreleased`
 
-- Add functionality to copy theme to clipboard.
 - Add reset functionality to revert changes.
 - Migrate state in Panel to Context.
+
+## `0.7.1` - 2020 July 27
+
+- Added functionality to copy changed theme to clipboard.
+- Added `CopyToClipboard` component.
 
 ## `0.7.0` - 2020 July 27